+++ js/ui/drive_lights.ts
@@ -9,11 +9,11 @@ export default class DriveLights implements Callbacks {
                 'url(css/red-off-16.png)';
     }
 
-    public dirty() {
+    public dirty(_drive: DriveNumber, _dirty: boolean) {
         // document.querySelector('#disksave' + drive).disabled = !dirty;
     }
 
-    public label(drive: DriveNumber, label: string) {
+    public label(drive: DriveNumber, label?: string) {
         const labelElement =
             document.querySelector('#disk-label' + drive)! as HTMLElement;
         if (label) {
